n&hash

+++ hash.h
@@ -13,7 +13,8 @@
 typedef uint32_t PPR;  // PRECISION OF THE POSITION POINTER FOR REPEATS
 
 typedef struct{
-  uint16_t key;      // THE KEY (INDEX / HASHSIZE) STORED IN THIS ENTRY
+//  uint16_t key;      // THE KEY (INDEX / HASHSIZE) STORED IN THIS ENTRY
+  uint64_t key;      // THE KEY STORED IN THIS ENTRY
   uint16_t nPos;
   PPR      *pos;     // LIST WITH THE REPEATING POSITIONS
   }
